requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
Django==3.1.2
django-allauth==0.42.0
idna==2.10
mysqlclient==2.0.1
oauthlib==3.1.0
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
sqlparse==0.4.1
urllib3==1.25.10

tabulate~=0.8.7
six~=1.15.0
websocket-client~=0.57.0
python-decouple==3.3
